some variables changes may be at hand
I might be wrong about this but when I try to use this extension now, it doesn't seem to work. So I went into typing.com developer tools to see if your functions matched what the typing elements were on the website, and it seems as if they just added "-line" or "-lines" to the end of all the variables. I dont have much experience in JavaScript or HTML. So maybe make some changes pls.